+++ OpenOrchestrator/orchestrator/tabs/logging_tab.py
@@ -11,9 +11,11 @@ from OpenOrchestrator.orchestrator import test_helper
 
 COLUMNS = [
     {'name': "Log Time", 'label': "Log Time", 'field': "Log Time", 'align': 'left', 'sortable': True},
-    {'name': "Process Name", 'label': "Process Name", 'field': "Process Name", 'align': 'left'},
     {'name': "Level", 'label': "Level", 'field': "Level", 'align': 'left'},
+    {'name': "Short Job ID", 'label': "Job ID", 'field': "Short Job ID", 'align': 'left', 'sortable': True},
+    {'name': "Process Name", 'label': "Process Name", 'field': "Process Name", 'align': 'left', 'sortable': True},
     {'name': "Message", 'label': "Message", 'field': "Message", 'align': 'left', ':format': 'value => value.length < 100 ? value : value.substring(0, 100)+"..."'},
+    {'name': "Full Job ID", 'label': "Full Job ID", 'field': "Full Job ID", 'headerClasses': 'hidden', 'classes': 'hidden'},
     {'name': "ID", 'label': "ID", 'field': "ID", 'headerClasses': 'hidden', 'classes': 'hidden'}
 ]
 
@@ -22,13 +24,18 @@ COLUMNS = [
 class LoggingTab():
     """The 'Logs' tab object."""
     def __init__(self, tab_name: str) -> None:
+        self.current_job_id: str | None = None
         with ui.tab_panel(tab_name):
-            with ui.row():
-                self.from_input = DatetimeInput("From Date", on_change=self.update, allow_empty=True)
-                self.to_input = DatetimeInput("To Date", on_change=self.update, allow_empty=True)
-                self.process_input = ui.select(["All"], label="Process Name", value="All", on_change=self.update).classes("w-48")
-                self.level_input = ui.select(["All", "Trace", "Info", "Error"], value="All", label="Level", on_change=self.update).classes("w-48")
-                self.limit_input = ui.select([100, 200, 500, 1000], value=100, label="Limit", on_change=self.update).classes("w-24")
+            with ui.row().classes("w-full justify-between"):
+                with ui.row():
+                    self.from_input = DatetimeInput("From Date", on_change=self.update, allow_empty=True)
+                    self.to_input = DatetimeInput("To Date", on_change=self.update, allow_empty=True)
+                    self.level_input = ui.select(["All", "Trace", "Info", "Error"], value="All", label="Level", on_change=self.update).classes("w-48")
+                    self.process_input = ui.select(["All"], label="Process Name", value="All", on_change=self.update).classes("w-48")
+                    self.limit_input = ui.select([100, 200, 500, 1000], value=100, label="Limit", on_change=self.update).classes("w-24")
+                with ui.column().classes("items-end") as self.job_filter_container:
+                    self.job_filter_label = ui.label("")
+                    self.all_jobs_button = ui.button("Show all jobs", on_click=self._show_all_jobs)
 
             self.logs_table = ui.table(title="Logs", columns=COLUMNS, rows=[], row_key='ID', pagination=50).classes("w-full")
             self.logs_table.on("rowClick", self._row_click)
@@ -39,16 +46,25 @@ class LoggingTab():
         """Update the logs table and Process input list"""
         self._update_table()
         self._update_process_input()
+        self._update_job_filter_display()
+
+    def _update_job_filter_display(self):
+        """Update visibility and text of job filter."""
+        if self.current_job_id:
+            self.job_filter_label.set_text(f"Filtering by Job: {self.current_job_id}...")
+            self.job_filter_container.set_visibility(True)
+        else:
+            self.job_filter_container.set_visibility(False)
 
     def _update_table(self):
         """Update the table with logs from the database applying the filters."""
         from_date = self.from_input.get_datetime()
         to_date = self.to_input.get_datetime()
-        process_name = self.process_input.value if self.process_input.value != 'All' else None
         level = LogLevel(self.level_input.value) if self.level_input.value != "All" else None
+        process_name = self.process_input.value if self.process_input.value != 'All' else None
         limit = self.limit_input.value
 
-        logs = db_util.get_logs(0, limit=limit, from_date=from_date, to_date=to_date, log_level=level, process_name=process_name)
+        logs = db_util.get_logs(0, limit=limit, from_date=from_date, to_date=to_date, log_level=level, process_name=process_name, job_id=self.current_job_id)
         self.logs_table.rows = [log.to_row_dict() for log in logs]
 
     def _update_process_input(self):
@@ -58,10 +74,18 @@ class LoggingTab():
         self.process_input.options = process_names
         self.process_input.update()
 
+    def set_job_filter(self, job_id: str | None):
+        """Set filter to specific job."""
+        self.current_job_id = job_id
+        self.update()
+
+    def _show_all_jobs(self):
+        self.set_job_filter(None)
+
     def _row_click(self, event):
         """Display a dialog with info on the clicked log."""
         row = event.args[1]
-        with ui.dialog(value=True), ui.card():
+        with ui.dialog(value=True) as dialog, ui.card():
             ui.label("Log ID:").classes("font-bold")
             ui.label(row['ID'])
             ui.label("Log Time:").classes("font-bold")
@@ -72,3 +96,8 @@ class LoggingTab():
             ui.label(row['Level'])
             ui.label("Message:").classes("font-bold")
             ui.html(f"<pre>{row['Message']}</pre>")
+            ui.label("Job ID:").classes("font-bold")
+            ui.label(row['Full Job ID'])
+            with ui.row():
+                ui.button("Show job logs", on_click=lambda: [self.set_job_filter(row['Full Job ID']), dialog.close()]).set_enabled(row['Full Job ID'] is not None)
+                ui.button("Close", on_click=dialog.close)