+++ OpenOrchestrator/tests/test_orchestrator_connection.py
@@ -2,7 +2,7 @@
 
 import unittest
 from datetime import datetime
-from uuid import UUID
+from uuid import UUID, uuid4
 import os
 
 from OpenOrchestrator.orchestrator_connection.connection import OrchestratorConnection
@@ -29,7 +29,7 @@ class TestOrchestratorConnection(unittest.TestCase):
                                                        is_blocking=False,
                                                        priority=0
                                                        )
-        cls.connection = OrchestratorConnection("Process", os.environ["CONN_STRING"], crypto_util.get_key(), "Args", cls.trigger_id)
+        cls.connection = OrchestratorConnection("Process", os.environ["CONN_STRING"], crypto_util.get_key(), "Args", str(cls.trigger_id), str(uuid4()))
 
     def test_trigger_pause(self):
         """Test pausing triggers."""