+++ OpenOrchestrator/tests/ui_tests/ui_util.py
@@ -35,7 +35,14 @@ def open_orchestrator() -> webdriver.Chrome:
     subprocess.Popen([sys.executable, "-m", "OpenOrchestrator", "o", "--port", str(port), "--dont_show"])  # pylint: disable=consider-using-with
 
     chrome_options = webdriver.ChromeOptions()
-    chrome_options.add_argument("--disable-search-engine-choice-screen")
+    custom_browser_options = [
+        '--disable-search-engine-choice-screen',
+        '--disable-dev-shm-usage',
+        '--log-level=3',
+        '--disable-logging'
+        ]
+    for option in custom_browser_options:
+        chrome_options.add_argument(option)
     browser = webdriver.Chrome(options=chrome_options)
     browser.implicitly_wait(2)
     browser.maximize_window()
@@ -54,6 +61,8 @@ def open_orchestrator() -> webdriver.Chrome:
     conn_input = browser.find_element(By.CSS_SELECTOR, "input[auto-id=connection_frame_conn_input]")
     conn_input.send_keys(Keys.CONTROL, "a", Keys.DELETE)
     conn_input.send_keys(conn_string)
+
+    time.sleep(1)
 
     browser.find_element(By.CSS_SELECTOR, "button[auto-id=settings_tab_key_button]").click()
     browser.find_element(By.CSS_SELECTOR, "button[auto-id=connection_frame_conn_button]").click()